Implement set unions.
The test added here could certainly use improving, thuogh I'm not sure
how. CPython seems to order sets some of the time, but not always, and
our behaviour is not the same yet. While this isn't an issue for the
sets themselves, which are inherently unoredered, it does make the tests
harder. The solution for the moment therefore is the internal comparison
I've used combined with careful printing of sets expected to be in the
right order to happen to match CPython.